students/brian_minsk/lesson07/test_html_render.py
"""
test code for html_render.py

This is just a start -- you will need more tests!
"""

import io

# import * is often bad form, but makes it easier to test everything in a module.
from html_render import *


# utility function for testing render methods
# needs to be used in multiple tests, so we write it once here.
# BSM: The requirement was to set the number of indent spaces manually as an
# attribute of the base class and not to be able to change it programmatically,
# so, for the Step 9 run, we will adapt the tests that fail with no
# indentation.
def render_result(element, ind=0):
    """
    calls the element's render method, and returns what got rendered as a
    string
    """
    # the StringIO object is a "file-like" object -- something that
    # provides the methods of a file, but keeps everything in memory
    # so it can be used to test code that writes to a file, without
    # having to actually write to disk.
    outfile = io.StringIO()
    # this so the tests will work before we tackle indentation
    if ind:
        # BSM: See my comment above.
        element.render(outfile)
    else:
        element.render(outfile)
    return outfile.getvalue()

########
# Step 1
########


def test_init():
    """
    This only tests that it can be initialized with and without
    some content -- but it's a start
    """
    e = Element()

    e = Element("this is some text")


def test_append():
    """
    This tests that you can append text

    It doesn't test if it works --
    that will be covered by the render test later
    """
    e = Element("this is some text")
    e.append("some more text")


def test_render_element():
    """
    Tests whether the Element can render two pieces of text
    So it is also testing that the append method works correctly.

    It is not testing whether indentation or line feeds are correct.
    """
    e = Element("this is some text")
    e.append("and this is some more text")

    # This uses the render_results utility above
    file_contents = render_result(e).strip()

    # making sure the content got in there.
    assert("this is some text") in file_contents
    assert("and this is some more text") in file_contents

    # make sure it's in the right order
    assert file_contents.index("this is") < file_contents.index("and this")

    # making sure the opening and closing tags are right.
    # This breaks with Step 9 so commenting out
    # assert file_contents.startswith("<html>")
    assert file_contents.endswith("</html>")


# Uncomment this one after you get the one above to pass
# Does it pass right away?
def test_render_element2():
    """
    Tests whether the Element can render two pieces of text
    So it is also testing that the append method works correctly.

    It is not testing whether indentation or line feeds are correct.
    """
    e = Element()
    e.append("this is some text")
    e.append("and this is some more text")

    # This uses the render_results utility above
    file_contents = render_result(e).strip()

    # making sure the content got in there.
    assert("this is some text") in file_contents
    assert("and this is some more text") in file_contents

    # make sure it's in the right order
    assert file_contents.index("this is") < file_contents.index("and this")

    # making sure the opening and closing tags are right.
    assert file_contents.startswith("<html>")
    assert file_contents.endswith("</html>")


# ########
# Step 2
# ########


# tests for the new tags
def test_html():
    e = Html("this is some text")
    e.append("and this is some more text")

    file_contents = render_result(e).strip()

    assert("this is some text") in file_contents
    assert("and this is some more text") in file_contents
    assert file_contents.endswith("</html>")


def test_body():
    e = Body("this is some text")
    e.append("and this is some more text")

    file_contents = render_result(e).strip()

    assert("this is some text") in file_contents
    assert("and this is some more text") in file_contents

    assert file_contents.startswith("<body>")
    assert file_contents.endswith("</body>")


def test_p():
    e = P("this is some text")
    e.append("and this is some more text")

    file_contents = render_result(e).strip()

    assert("this is some text") in file_contents
    assert("and this is some more text") in file_contents

    assert file_contents.startswith("<p>")
    assert file_contents.endswith("</p>")


def test_sub_element():
    """
    tests that you can add another element and still render properly
    """
    page = Html()
    page.append("some plain text.")
    page.append(P("A simple paragraph of text"))
    page.append("Some more plain text.")

    file_contents = render_result(page)
    print(file_contents)  # so we can see it if the test fails

    # note: The previous tests should make sure that the tags are getting
    #       properly rendered, so we don't need to test that here.
    assert "some plain text" in file_contents
    assert "A simple paragraph of text" in file_contents
    assert "Some more plain text." in file_contents
    assert "some plain text" in file_contents
    # but make sure the embedded element's tags get rendered!
    assert "<p>" in file_contents
    assert "</p>" in file_contents


########
# Step 3
########

# Run all the Step 2 tests - don't comment them out for Step 3

def test_head():
    """ Test that the head tag renders properly.
    """

    e = Head("this is some text")
    e.append("and this is some more text")

    file_contents = render_result(e).strip()

    assert("this is some text") in file_contents
    assert("and this is some more text") in file_contents

    assert file_contents.startswith("<head>")
    assert file_contents.endswith("</head>")


def test_title():
    """ Test that the title tag renders properly.
    """

    e = Title("this is some text")
    e.append("and this is some more text")

    file_contents = render_result(e).strip()

    assert("this is some text") in file_contents
    assert("and this is some more text") in file_contents

    assert file_contents.startswith("<title>")
    assert file_contents.endswith("</title>")


def test_one_line_tag():
    """ Test to see if the tags and text between the tags are on the same line.
    Title is a subclass of OneLineTag.
    """
    e = Title("this is some text")
    e.append("and this is some more text")

    file_contents = render_result(e).strip()

    assert("\n") not in file_contents
    assert("\r") not in file_contents


########
# Step 4
########

# Run all the Step 2 & 3 tests - don't comment them out for Step 4

def test_one_tag_attribute():
    """ Test that a single tag attribute renders properly.
    """

    e = Html()
    e.append(P("Here is a paragraph of text -- there could be more of them, "
               "but this is enough  to show that we can do some text",
                style="text-align: center; font-style: oblique;"))

    file_contents = render_result(e).strip()

    assert('<p style="text-align: center; font-style: oblique;">')in file_contents


def test_two_tag_attributes():
    """ Test that two tag attributes render properly.
    """

    e = Html()
    e.append(P("Here is a paragraph of text -- there could be more of them, "
               "but this is enough  to show that we can do some text",
                style="text-align: center; font-style: oblique;", spellcheck="true"))

    file_contents = render_result(e).strip()

    assert('<p style="text-align: center; font-style: oblique;" spellcheck="true">') in file_contents

########
# Step 5
########

# Run all the Step 2, 3, & 4 tests - don't comment them out for Step 5


def test_hr():
    """ Test that the hr tag renders properly.
    """

    e = Html()
    e.append(Hr())

    file_contents = render_result(e).strip()

    assert('<hr />') in file_contents


def test_hr_with_attr():
    """ Test that the hr tag with an attribute renders properly.
    """

    e = Html()
    e.append(Hr(display="block", title="seperator"))

    file_contents = render_result(e).strip()

    assert('<hr display="block" title="seperator" />') in file_contents


def test_br():
    """ Test that the br tag renders properly.
    """

    e = Html()
    e.append(Br())

    file_contents = render_result(e).strip()

    assert('<br />') in file_contents


########
# Step 6
########

# Run all the Step 2, 3, 4, & 5 tests - don't comment them out for Step 6

def test_init_a():
    """
    This only tests that it can be initialized with and without
    some content -- but it's a start
    """
    a = A("https://cat-bounce.com/", "CAT BOUNCE!")


def test_a():
    e = Html()
    e.append(A("http://google.com", "link to google"))

    file_contents = render_result(e).strip()

    assert('<a href="http://google.com">link to google</a>') in file_contents


########
# Step 7
########

# Run all the Step 2, 3, 4, 5, & 6 tests - don't comment them out for Step 7

def test_init_h():
    """
    This only tests that it can be initialized with and without
    some content -- but it's a start
    """
    h = H(2)

    h = H(3, "some text", some_kwarg="value")


def test_h():
    """ Test that the h tag renders properly.
    """

    e = Html()
    e.append(H(1, "This is level 1 header"))
    e.append(H(2, "This is level 2 header"))
    e.append(H(3, "This is level 3 header"))

    file_contents = render_result(e).strip()

    assert('<h1>This is level 1 header</h1>') in file_contents
    assert('<h2>This is level 2 header</h2>') in file_contents
    assert('<h3>This is level 3 header</h3>') in file_contents


def test_ul():
    """ Test that the ul tag renders properly.
    """

    e = Html()
    e.append(Ul(id="AList", style="line-height:150%"))

    file_contents = render_result(e).strip()

    assert('<ul id="AList" style="line-height:150%">') in file_contents
    assert('</ul>') in file_contents


def test_li():
    """ Test that the li tag renders properly.
    """

    e = Html()
    e.append(Li("An item in a list"))
    
    file_contents = render_result(e).strip()

    lines = file_contents.split("\n")

    assert lines[2] == (" " * e.indent) + '<li>'
    assert lines[3] == (" " * 2 * e.indent) + 'An item in a list'
    assert lines[4] == (" " * e.indent) + '</li>'


def test_li_with_attr():
    """ Test that the li tag with an attribute renders properly.
    """

    e = Html()
    e.append(Li("An item in a list",  style="color: blue"))

    file_contents = render_result(e).strip()

    lines = file_contents.split("\n")

    assert lines[2] == (" " * e.indent) + '<li style="color: blue">'
    assert lines[3] == (" " * 2 * e.indent) + 'An item in a list'
    assert lines[4] == (" " * e.indent) + '</li>'


########
# Step 8
########

# Run all the Step 2, 3, 4, 5, 6, 7 tests - don't comment them out for Step 8

def test_doctype_tag():
    """ Test that the doctype tag gets rendered properly.
    """

    e = Html()

    file_contents = render_result(e).strip()

    assert file_contents.startswith('<!DOCTYPE html>')


def test_meta():
    """ Test that the meta tag renders properly.
    """

    e = Html()
    head = Head()
    e.append(head)
    head.append(Meta(charset="UTF-8"))

    file_contents = render_result(e).strip()

    assert('<meta charset="UTF-8" />') in file_contents


# #####################
# indentation testing
# Uncomment for Step 9 -- adding indentation
# #####################


# BSM: I don't understand this test because the <html> tag shouldn't be indented
# Also the first line - the one preceding the <html> tag should be <!DOCTYPE html>
# I've written a new test_indent below.
# def test_indent():
#     """
#     Tests that the indentation gets passed through to the renderer
#     """
#     html = Html("some content")
#     # Following doesn't work since my Element.indent is an int
#     # file_contents = render_result(html, ind="   ").rstrip()  #remove the end newline
#     # Changed to this:
#     file_contents = render_result(html, ind=3).rstrip()  #remove the end newline

#     print(file_contents)
#     lines = file_contents.split("\n")
#     assert lines[0].startswith("   <")
#     print(repr(lines[-1]))
#     assert lines[-1].startswith("   <")

def test_indent():
    """
    Tests that the indentation gets passed through to the renderer
    """
    html = Html()
    body = Body("some content")
    html.append(body)
    # Following doesn't work since my Element.indent is an int
    # file_contents = render_result(html, ind="   ").rstrip()  #remove the end newline
    # Changed to this:
    file_contents = render_result(html, ind=3).rstrip()  # remove the end newline

    print(file_contents)
    lines = file_contents.split("\n")
    assert lines[2].startswith("   <")
    print(repr(lines[-2]))
    assert lines[-2].startswith("   <")


def test_indent_contents():
    """
    The contents in a element should be indented more than the tag
    by the amount in the indent class attribute
    """
    html = Element("some content")
    file_contents = render_result(html, ind="")

    print(file_contents)
    lines = file_contents.split("\n")
    # Following doesn't work since my Element.indent is an int
    # assert lines[1].startswith(Element.indent)
    # Changed to this:
    assert lines[1].startswith(" " * Element.indent)


def test_multiple_indent():
    """
    make sure multiple levels get indented fully
    """
    body = Body()
    body.append(P("some text"))
    html = Html(body)

    file_contents = render_result(html)

    print(file_contents)
    lines = file_contents.split("\n")
    for i in range(3):  # this needed to be adapted to the <DOCTYPE> tag
        # Following doesn't work since my Element.indent is an int
        # assert lines[i + 1].startswith(i * Element.indent + "<")
        # Changed to this:
        assert lines[i + 1].startswith(i * (" " * Element.indent) + "<")

    # Following doesn't work since my Element.indent is an int
    # assert lines[4].startswith(3 * Element.indent + "some")
    # Changed to this:
    assert lines[4].startswith(3 * (" " * Element.indent) + "some")


def test_element_indent1():
    """
    Tests whether the Element indents at least simple content

    we are expecting to to look like this:

    <html>
        this is some text
    </html>

    More complex indentation should be tested later.
    """
    e = Element("this is some text")

    # This uses the render_results utility above
    file_contents = render_result(e).strip()

    # making sure the content got in there.
    assert("this is some text") in file_contents

    # break into lines to check indentation
    lines = file_contents.split('\n')
    # making sure the opening and closing tags are right.
    assert lines[0] == "<html>"
    # this line should be indented by the amount specified
    # by the class attribute: "indent"
    # Since I used an int for Element.indent the following doesn't work
    # assert lines[1].startswith(Element.indent + "thi")
    assert lines[1].startswith((" " * Element.indent) + "thi")  # changed previous line to this
    assert lines[2] == "</html>"
    assert file_contents.endswith("</html>")


def test_element_indent2():
    """
    Tests two levels of element indents

    we are expecting to to look like this:

    <html>
        <b>
            this is some text
        </b>
    </html>

    """
    e = Element()
    b = Body("this is some text")
    e.append(b)

    # This uses the render_results utility above
    file_contents = render_result(e).strip()

    # making sure the content got in there.
    assert("this is some text") in file_contents
    print(file_contents)

    # break into lines to check indentation
    lines = file_contents.split('\n')
    # making sure the opening and closing tags are right.
    assert lines[0] == "<html>"
    assert lines[-1] == "</html>"
    # Make sure the 2nd level opening and closing tags are right.
    # These lines should be indented by the amount specified
    # by the class attribute: "indent"
    assert lines[1] == (" " * Element.indent) + "<body>"
    assert lines[-2] == (" " * Element.indent) + "</body>"

    # Make sure the 3rd level indent is right.
    # This line should be indented by twice the amount specified
    # by the class attribute: "indent"
    assert lines[2] == (" " * 2 * Element.indent) + "this is some text"
